rename UserList dir
